Move `FeedImageCell` to its own file
